fix bug and tests
