--- packages/common/src/dcx-trusted-issuers.ts
@@ -1,9 +1,0 @@
-export const MX = {
-  name : 'mx',
-  id   : 'did:dht:sa713dw7jyg44ejwcdf8iqcseh7jcz51wj6fjxbooj41ipeg76eo'
-};
-export const FORMFREE = {
-  name : 'formfree',
-  id   : 'did:dht:hcf5e55bbm44s4oixp5z89wtxenxyk35su7f5pd4r5np93ikyowy'
-};
-export const defaultTrustedIssuers = [MX, FORMFREE];
