+++ packages/common/src/index.ts
@@ -12,7 +12,6 @@ export { schema as manifestSchema } from './schemas/manifest.js';
 export type * from './types/did.js';
 export type * from './types/handlers.js';
 export type * from './types/options.js';
-export type * from './types/config.js';
 export type * from './types/pex.js';
 export type * from './types/server.js';
 export type * from './types/web5.js';
@@ -36,4 +35,3 @@ export * from './dcx-config.js';
 export * from './dcx-dht-manager.js';
 export * from './dcx-identity-vault.js';
 export * from './dcx-manager.js';
-export * from './dcx-trusted-issuers.js';
